Remove `bytes` dependency
With the release of `tokio@1` and `bytes@1`, `rusoto_dynamodb` is likely
to update to a newer version soon.

When that happens, we will want to pull in the version of `bytes` that
corresponds to the version that `rusoto_dynamodb` is using. But here's
the thing: cargo doesn't really let us do that, and `rusoto_dynamodb`
doesn't expose `bytes` in any way we can get at.

But what we _really_ need is to be able to create whatever type is an
`AttributeValue.b` from a slice of bytes, and turn whatever type is an
`AttributeValue.b` into a slice of bytes. And fortunately, `bytes`
implements both of this generically, as `From<Vec<u8>>` and
`AsRef<[u8]>`.

So instead of naming bytes specifically, if we just deal with types that
implement `From<Vec<u8>>` and `AsRef<[u8]>` generically, we can support
whichever version of `bytes` our key `rusoto_dynamodb` dependency needs
transparently. 🥳