aws: Update docs for `.items` → `.items()`
AWS SDK for DynamoDB used to expose the `.items` field from [ScanOutput],
etc.

As of version 0.16.0, the documentation for `.items` is hidden,
suggesting that the best way to interact with this structure is to use
`.items()` to get a slice of the items, and unnecessarily clone that
slice if you want to use it.

To prevent confusion, update the docs to use `.items()` instead of
`.items`.

Fixes #54, #55

[ScanOutput]: https://docs.rs/aws-sdk-dynamodb/0.22.0/aws_sdk_dynamodb/output/struct.ScanOutput.html