Fix for aws-sdk 0.35 interface change
aws-sdk-dynamo's ScanOutput and QueryOutput changed slightly in the most
recent release: 0.35.

Previously

    pub fn items(&self) -> Option<&[HashMap<String, AttributeValue>]>

Now

    pub fn items(&self) -> &[HashMap<String, AttributeValue>]

While this doesn't affect serde_dynamo directly, we do have a bunch of
doc comments that use `result.items()`. Update those so that the tests
pass again.

This was missed on the 0.35 PR because tests had been failing for
another reason (#80) for a while, so I didn't even bother looking at the
failing tests.